Restructure theories
Splits the sorts out of the large terms.rs into separate modules for
each theory.

There is still work yet to be done on all of the theories.

For BitVec the current implementation still seems to have some issues
that needs to be fixed.

It seems that many of the functions for theories are not explicitly
defined in the theories, but rather are states as part of a logic.
For an example of this, check out FixedBitVectors.smt2 and QF_BV.smt2.