Add a z3 statically linked backend using z3 crate
The communication layer uses the Z3_eval_smtlib2_string in order to send
the formatted SMT-LIB strings to, instead for constructing an Z3 AST
using their API.

This has the benefit of working pretty much like a CLI binary, and thus
fits right with the other backends!

Some non-scientific performance measurements show that this backend is
not as performant as one would hope, and gets outperformed by the z3 CLI
by quite a bit! This might be interesting to look into at some point.
But for now, the CLI binary interfaces win WRT performance!

In the current implementation, the communication requires a handful of
unsafe calls. This is because the Z3_eval_smtlib2_string API is not
exposed in the higher-level z3 crate and we thus has to rely on the
z3-sys crate :/