feat!: Remove `Async{Driver,Solver}` in favor of `Tokio{Driver,Solver}`
The previous implementations simply created a future and ran blocking code in that future. This was pretty much all we could do without having an actual variant of `Command` that was async.

So instead of trying to find some executor-agnostic approach, we will be explicit about using tokio.

This explicitness will allow is in the future to add other executors, and not give a false impression about being agnostic.

We might lift this or change the naming in the future.

One thing to note, is that the currenting naming is a bit strange; sort of a mix of suffixing/prefix `Tokio` where it fits. I could not immediately come up with a holistic way of doing this, so we stick with this for now.